+++ Lib-UAV/src/lib/uav/struct/constants/TypeFailureInsert.java
@@ -1,0 +1,15 @@
+package lib.uav.struct.constants;
+
+/**
+ * The class models the types of failure insert.
+ * @author Jesimar S. Arantes
+ * @since version 5.0.0
+ */
+public class TypeFailureInsert {
+    
+    public static final String NONE     = "NONE";
+    public static final String WAYPOINT = "WAYPOINT";
+    public static final String TIME     = "TIME";
+    public static final String POSITION = "POSITION";
+    
+}